refactor: controller에 있는 검증 메서드 분리
